:recycle: refactor: move solution adapter to src
