+++ api/tests/unit/infrastructure/adapters/solution-adapter_test.js
@@ -1,5 +1,5 @@
 import { expect, domainBuilder } from '../../../test-helper.js';
-import * as solutionAdapter from '../../../../lib/infrastructure/adapters/solution-adapter.js';
+import * as solutionAdapter from '../../../../src/shared/infrastructure/adapters/solution-adapter.js';
 import { Solution } from '../../../../lib/domain/models/Solution.js';
 
 describe('Unit | Adapter | Solution', function () {
